% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnp.R
\docType{data}
\name{bnp}
\alias{bnp}
\title{Data: Biscayne Bay polygon file}
\format{A SpatialPolygonDataFrame of Biscayne Bay.
\describe{
\item{bnp}{a shapefile of Biscayne Bay}
}}
\usage{
bnp
}
\description{
Data: Biscayne Bay polygon file
}
\examples{
plot(bnp)

}
\keyword{bay,}
\keyword{biscayne}
\keyword{map}
