% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpMod.R
\docType{data}
\name{bnpMod}
\alias{bnpMod}
\title{Data: Biscayne Bay model boundaries}
\format{A shapefile of model boundaries in Biscayne Bay, Florida. Coordinate reference system is "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
\describe{
\item{HECTARES}{area of basin in hectares}
\item{BOX_CODE}{basin code}
\item{Box_Basin}{name of basin}
}}
\usage{
bnpMod
}
\description{
Data: Biscayne Bay model boundaries
}
\examples{
plot(bnpMod)

}
\keyword{bay,}
\keyword{biscayne}
\keyword{map}
