% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalculationsParallel.R
\name{ciCalculationsParallel}
\alias{ciCalculationsParallel}
\title{A parallel-processing version fo EGRETci's bootAnnual}
\usage{
ciCalculationsParallel(eList, probs, clusterObject, nBoot = 100, 
    blockLength = 200, widthCI = 90, seed_var = 23, ...)
}
\arguments{
\item{eList}{eList object returned from EGRET's WRTDS}

\item{probs}{a vector of probabilities}

\item{clusterObject}{name of cluster object}

\item{nBoot}{number of bootstrap replicates to run}

\item{blockLength}{length of block used to identify neighboring water quality samples}

\item{widthCI}{width of confidence interval (percentiles)}

\item{seed_var}{seed to use for replicable results}

\item{...}{placeholder. Additional    arguments are not used at this time,}
}
\value{
list \code{ciCalculationsParallel} returns a list of dataframes
}
\description{
Runs EGRETci::bootAnnual in parallel.
}
\examples{
\dontrun{
registerDoParallel(cl)
nitro <- lapply(targStns, function(stnSelect)  
  modelEstimation(convertToEgret(stn = stnSelect, target_analyte = targAnalyte,  
                                   wq_data = wqTemp, flow_data = hydDat)))
                                   stopCluster(cl)
nitro <- lapply(nitro, setPA, paStart = WaterYearStart, paLong = 12)

caseSetUp.tkn <- mapLists(trendSetUp, nitro, list2 = NULL, 
    year2 = list(2017, 2017, 2015, 2015, 2007, 2013), year1 = startDate, 
    nBoot = nBoot_var, min = 100, blockLength = blockLength_var, bootBreak = 100)

eBoot.tkn <- mapLists(wBT, nitro, caseSetUp.tkn)

closeAllConnections()
CIAnnualResults.tkn <- lapply(nitro, ciCalculations, nBoot = nBoot_CI, 
    blockLength = blockLength_var, widthCI = 90)
closeAllConnections()
}

}
