% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInfo.R
\name{createInfo}
\alias{createInfo}
\title{Create 'INFO' metadata for EGRET/WQTDS analysis}
\usage{
createInfo(wq_data, paStart = 10, paLong = 12, watershedKm = 1)
}
\arguments{
\item{wq_data}{a dataframe of water quality data generated by getwq(). Must have a single station and parameter.}

\item{paStart}{Starting month of period of analysis. Defaults to 10. Used in most EGRET functions}

\item{paLong}{Length in number of months of period of analysis. Defaults to 12. Used in most EGRET functions}

\item{watershedKm}{Watershed area, used to calculate runoff. Defaults to 1}
}
\value{
a single-row dataframe with EGRET-style metadata. See ?EGRET::INFOdataframe for more information.
}
\description{
Create 'INFO' metadata for EGRET/WQTDS analysis
}
\examples{
INFO <- createInfo(wq_data = wqDat)
}
