% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_proj.R
\name{dbhydro.proj}
\alias{dbhydro.proj}
\title{An API for project-level DBHYDRO water quality data}
\usage{
dbhydro.proj(destfile = "proj_report_todaysDate.csv",
project_codes = "ACMEB", 
start_date = "01-JAN-1960", # note format
destination = "file_csv", 
rename_proj = TRUE,
import_data = FALSE)
}
\arguments{
\item{destfile}{a character vector file address for output, saved as a comma-delimited text file.}

\item{project_codes}{a character vector of station names. If one station in the list returns no data, the entire url request fails.}

\item{start_date}{lorem ipsum}

\item{destination}{lorem ipsum}

\item{rename_proj}{indicates whether destination file be re-named to replace "proj" with project name}

\item{import_data}{lorem ipsum}
}
\value{
dataframe \code{dbhydro.proj} saves a csv of DBHYDRO water quality data to disk and has an option (\code{import = TRUE}) to load a dataframe of data into working environment.
}
\description{
Downloads DBHYDRO water quality data for a specified project
}
\examples{
\dontrun{
# save data as a csv:
dbhydro.proj(project_codes = "CAMB")

# or load into working environment
dbhydro.proj(project_codes = "CAMB", import = TRUE)
}

}
\seealso{
\code{\link{dbhydro.proj.batch}} is preferred for multiple-project downloads.
}
