% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_stn_batch.R
\name{dbhydro.stn.batch}
\alias{dbhydro.stn.batch}
\title{A wrapper for \code{dbhydro.stn}}
\usage{
dbhydro.stn.batch(codes, ...)
}
\arguments{
\item{codes}{a character vector of station names. If one station in the list returns no data, the entire url request fails.}

\item{...}{additional arguments supplied to \code{\link{dbhydro.stn}}}
}
\value{
dataframe \code{dbhydro.stn.batch} saves a csv of DBHYDRO water quality data to disk.
}
\description{
Downloads DBHYDRO water quality data for specified stations. Avoids the problem of complete failure if a single station is invalid.
}
\examples{
\dontrun{
head(dbh.stn.batch(codes = c("S333", "S151"))
}

}
\seealso{
\code{\link{dbhydro.stn}}
}
