% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_stn.R
\name{dbhydro.stn}
\alias{dbhydro.stn}
\title{An API for station-level DBHYDRO water quality data}
\usage{
dbhydro.stn(destfile = "stn_report_todaysDate.csv",
stations = c("ACRA1"),
rename_proj = TRUE,
parameters = "all",
report_type = "full", # full = long dataset - 1 line per sample; crosstab = wide dataset
incl_qc_flags = TRUE,
incl_flagged_data = FALSE,
destination = "file_csv", 
start_date = "01-JAN-1960", # note format
end_date   = "today",
import_data = FALSE)
}
\arguments{
\item{destfile}{a character vector file address for output, saved as a comma-delimited text file.}

\item{stations}{a character vector of station names. If one station in the list returns no data, the entire url request fails.}

\item{rename_proj}{lorem ipsum}

\item{parameters}{lorem ipsum}

\item{report_type}{lorem ipsum}

\item{incl_qc_flags}{lorem ipsum}

\item{incl_flagged_data}{lorem ipsum}

\item{destination}{lorem ipsum}

\item{start_date}{lorem ipsum}

\item{end_date}{lorem ipsum}

\item{import_data}{lorem ipsum}
}
\value{
dataframe \code{dbhydro.stn} saves a csv of DBHYDRO water quality data to disk.
}
\description{
Downloads DBHYDRO water quality data for specified stations
}
\examples{
\dontrun{
head(dbhydro.stn())
}

}
\seealso{
\code{\link{dbhydro.stn.batch}} is preferred for multiple-station downloads.
}
