% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHYDROToDFE.R
\name{DBHYDROToDFE}
\alias{DBHYDROToDFE}
\title{Convert DBHYDRO flow and water quality datasets to match DataForEver for use in convertToEgret()}
\usage{
DBHYDROToDFE(data, type = "hydrology")
}
\arguments{
\item{data}{input dataset from DBHYDRO. Dataset will be converted to be consistent with data from DataForEver.}

\item{type}{type of input data: "hydrology" or "waterquality"}
}
\value{
a list as created by \code{\link[EGRET]{mergeReport}}
}
\description{
Convert DBHYDRO flow and water quality datasets to match DataForEver for use in convertToEgret()
}
\examples{
targStn <- "S333"
}
