% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_CQ.R
\name{dfe.CQ}
\alias{dfe.CQ}
\title{Plot concentration-discharge relationship for a DataForEver station}
\usage{
dfe.CQ(stn, parameter_list = c("flow", "tail_water", "head_water", "stage"),
target_analytes = "orthophosphate", date_range = "all", rFriendlyParamNames = TRUE
)
}
\arguments{
\item{stn}{the target station from DataForEver. Recommended to run function on a single station at a time.}

\item{parameter_list}{vector of desired parameters from the hydrology database}

\item{target_analytes}{grep-style character vector naming water quality analytes of interest. default is "orthophosphate". Case insensitive.}

\item{date_range}{date range to be included. Default is entire period of record.}

\item{rFriendlyParamNames}{TRUE/FALSE; indicates whether parameter names should be modified to be R-friendly (no special characters, commas, or spaces). Advisable for analysis, as this makes analysis easier and pre-empts changes coerced by, e.g., \code{plyr::ddply}}
}
\value{
list \code{dfe.CQ} returns a plot of the CQ dataframe with water quality measurements from each station identified in \code{stns}.
}
\description{
Downloads and compiles DataForEver hydrology and water quality data. Creates plot and returns descriptive statistics and raw data in list form. This function Works only on linux machines on the SFNRC network with access to the opt/physical drive. Code issues system commands, runs shell scripts, and modifies files in a temp folder on the local hard drive.
}
\examples{
\dontrun{
dfe.CQ(stns = "S700")
}

}
\seealso{
\code{\link{dfe.wq}}, \code{\link{dfe.hydro}}
}
