% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_data_types.R
\name{dfe.data.types}
\alias{dfe.data.types}
\title{DataForEver hydrology database parameter query}
\usage{
dfe.data.types(query = "flow")
}
\arguments{
\item{query}{a character string}
}
\value{
dataframe \code{dfe.data.types} returns a vector of stations
}
\description{
Identifies parameters available from the DataForEver hydrology database
}
\examples{
a <- dfe.data.types(query = "flow")

}
