% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_stn.R
\name{dfe.stn}
\alias{dfe.stn}
\title{DataForEver station query}
\usage{
dfe.stn(query = "S33")
}
\arguments{
\item{query}{a character string. Case insensitive.}
}
\value{
dataframe \code{dfe.stn} returns a vector of stations
}
\description{
Identifies DataForEver stations matching query
}
\examples{
a <- dfe.stn(query = "S33")
a <- dfe.stn(query = "s333")

}
