% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_wq.R
\name{dfe.wq}
\alias{dfe.wq}
\title{DataForEver water quality database API}
\usage{
dfe.wq(stns, target_analytes = "all", 
output_colNames = c("stn", "date", "time", "param", "units", 
"matrix", "mdl", "value"),
output_colClasses = c("character", "character", "character", 
"character", "character", "character", "numeric", "numeric"))
}
\arguments{
\item{stns}{a single-column dataframe of DataForEver station names.}

\item{target_analytes}{grep-style character vector naming analytes of interest. default is "all". e.g., # target_analytes <- c("PHOSPHATE|NITROGEN|AMMONI|SUSPENDED|DISSOLVED OXYGEN|CALCIUM|POTASSIUM|HARDNESS|SODIUM|CHLORIDE|TEMP|CONDUCTIVITY, FIELD|SILICA|LEAD, TOTAL|MAGNESIUM|TURBIDITY|CHLOROPHYLL|MERCURY, TOTAL|SULFATE|ZINC, TOTAL|CHLORDANE|MALATHION|CARBOPHENOTHION|PH, FIELD")}

\item{output_colNames}{names of columns in output (changing this is not recommended; this argument is more of a placeholder)}

\item{output_colClasses}{classes of columns in output (changing this is not recommended; this argument is more of a placeholder)}
}
\value{
dataframe \code{dfe.wq} returns a dataframe with water quality measurements from each station identified in \code{stns}.
}
\description{
Downloads and compiles DataForEver water quality data.
}
\details{
This function Works only on linux machines on the SFNRC network with access to the opt/physical drive. Code issues system commands, runs shell scripts, and modifies files in a temp folder on the local hard drive.
}
\examples{
\dontrun{
stations <- c("S333", "S12A", "S12B", "S12C", "S12D")

head(dfe.wq(stns = stations, target_analytes = c("PHOSPHATE|NITROGEN|AMMONI")))
}

}
\seealso{
\code{\link{dfe.hydro}}
}
