% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNWS.R
\name{downloadNWS}
\alias{downloadNWS}
\title{Function bulk downloads National Weather Service precipitation geotiff files}
\usage{
downloadNWS(
  data_folder,
  overwrite_data = FALSE,
  stage = c("stage3", "stage4"),
  stage3_filename = "nws_precip_mtd_YYYYMMDD.tif",
  stage4_filename = "nws_precip_mtd_YYYYMMDD_conus.tif",
  day_range = 1
)
}
\arguments{
\item{data_folder}{parent directory for data to be deposited. If this does not exist, it will be created.}

\item{overwrite_data}{TRUE/FALSE value indicating whether existing files should be overwritten. If TRUE, all data are downloaded fresh.}

\item{stage}{type of data to download. Acceptable values include `stage3` and/or `stage4`.}

\item{stage3_filename}{example of a stage 3 filename. YYYYMMDD is a token that will be replaced with actual dates. Adjusting this can accommodate odd filenames (e.g., stage 3 2017; stage 4 2016) and download different data types (e.g., 1 day, last 180 days)}

\item{stage4_filename}{example of a stage 4 filename. YYYYMMDD is a token that will be replaced with actual dates.}

\item{day_range}{numeric range of days to download from each month. Default only downloads data from the first of the month.}
}
\value{
NULL \code{downloadNWS} downloads files to the provided `data_folder` directory.
}
\description{
Useful for automating download of large numbers of NWS spatial precipitation layers, and easily download new data. Stage 3 and Stage 4 data can be found at https://water.noaa.gov/resources/downloads/precip/
}
\examples{
2+2
}
