% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBootReport.R
\name{eBootReport}
\alias{eBootReport}
\title{Extract useful information from bootstrapped WRTDS output}
\usage{
eBootReport(data = eBoot, eList = NULL,
flux = FALSE, conc = TRUE, text = FALSE)
}
\arguments{
\item{data}{eBoot results from EGRETci}

\item{eList}{a list (of data frames, models, etc.)}

\item{flux}{logical, indicating whether output is desired for fluxes}

\item{conc}{logical, indicating whether output is desired for concenctrations}

\item{text}{logical, indicates whether text output should be printed to the console}
}
\value{
list \code{eBootReport} returns a list the same length as list1 and list2
}
\description{
Summarizes concentration and/or flux trends identified by EGRETci utilities.
}
\examples{
\dontrun{
registerDoParallel(cl)
nitro <- lapply(targStns, function(stnSelect)  
  modelEstimation(convertToEgret(stn = stnSelect, target_analyte = targAnalyte,  
                                   wq_data = wqTemp, flow_data = hydDat)))
                                   stopCluster(cl)
nitro <- lapply(nitro, setPA, paStart = WaterYearStart, paLong = 12)

caseSetUp.tkn <- mapLists(trendSetUp, nitro, list2 = NULL, 
    year2 = list(2017, 2017, 2015, 2015, 2007, 2013), year1 = startDate, 
    nBoot = nBoot_var, min = 100, blockLength = blockLength_var, bootBreak = 100)

eBoot.tkn <- mapLists(wBT, nitro, caseSetUp.tkn)

eBootReport(data = eBoot.tkn[[1]]$pConc)
}
 
}
