% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMean.R
\name{geoMean}
\alias{geoMean}
\title{Calculate geometric mean}
\usage{
geoMean(x, nas = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{vector of values from which geometric mean will be calculated}

\item{nas}{TRUE/FALSE value indicating whether NAs should be removed}

\item{zero.propagate}{TRUE/FALSE value indicating whether zeroes should be propagated}
}
\value{
vector \code{geoMean} returns the geometric mean of the input values
}
\description{
Calculates the geometric mean of a vector of values.
}
\examples{
geoMean(rnorm(100, 5, 5))
}
