% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordsEDEN.R
\name{getCoords_EDEN}
\alias{getCoords_EDEN}
\title{Retrieve station coordinates from EDEN}
\usage{
getCoords_EDEN(stn = "S18C_T", spatial = TRUE)
}
\arguments{
\item{stn}{character string. Case sensitive.}

\item{spatial}{logical. If `TRUE`, a spatialPointsDataFrame is returned. If `FALSE`, a dataframe is returned}
}
\value{
dataframe/spdf \code{getCoords_EDEN} returns a dataframe or spatialPointsDataFrame with lat/long (epsg:4326) UTM zone 19 coordinates, and the NAVD-to-NGVD conversion factor (units = feet). [ft. NAVD88] + [conversion factor] = [ft. NGVD29]
}
\description{
Scrapes EDEN station coordinate data.
}
\examples{
\dontrun{
### example usage as part of a workflow

stns <- c(
"S18C_T",
"NESRS1"
)
stn.coords <- do.call(rbind, lapply(X = stns, getCoords_EDEN))


plot(stn.coords)
}

}
