% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_data_types.R
\name{getDataTypes}
\alias{getDataTypes}
\title{DataForEver hydrology database parameter query}
\usage{
getDataTypes(parameter = "all", stn = "all", exactMatch = FALSE)
}
\arguments{
\item{parameter}{a character string specifying the parameter(s) used to constrain the query.}

\item{stn}{a character string specifying the station(s) used to constrain the query.}

\item{exactMatch}{If \code{TRUE}, station name must be an exact match with the \code{stn} argument}
}
\value{
dataframe \code{getDataTypes} returns a vector of stations and parameters.
}
\description{
Identifies parameters available from the DataForEver hydrology database
}
\examples{
\dontrun{
### search by parameter:
getDataTypes(parameter = "salinity")

### search by station:
getDataTypes(stn = "S333")
getDataTypes(stn = "S333", exactMatch = TRUE) # note that the function does not use exact matches
     # unless requested to do so using \code{fixed = TRUE}
}


}
