% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBHYDROparams.R
\name{getDBHYDROparams}
\alias{getDBHYDROparams}
\title{Reports available water quality parameters from DBHYDRO}
\usage{
getDBHYDROparams(stn = "S333", parameters = NA)
}
\arguments{
\item{stn}{character string. Case insensitive.}

\item{parameters}{a grep-style character string identifying water quality parameters to report. Set to NA to report all available parameters. Partial matches are made (e.g., "TURB" and "TURBIDITY" will both yield turbidity data)}
}
\value{
dataframe \code{getDBHYDROparams} returns a vector of parameter names and the number of observations for each.
}
\description{
a simple download-and-query, designed to get parameter names correct in future downloads
}
\examples{
\dontrun{
tableDat <- getDBHYDROparams(stn = "s333") # all parameters
tableDat

phosDat <- getDBHYDROparams(stn = "s333", parameters = "PHOS") # using parameter filter
phosDat
}

 
}
