% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_stn.R
\name{getStn}
\alias{getStn}
\title{DataForEver station query}
\usage{
getStn(query = "S33")
}
\arguments{
\item{query}{a character string. Case insensitive.}
}
\value{
dataframe \code{getStn} returns a vector of stations
}
\description{
Identifies DataForEver stations matching query
}
\examples{
a <- getStn(query = "S33")
a <- getStn(query = "s333")

}
