% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_wq.R
\name{getWQ}
\alias{getWQ}
\title{DataForEver water quality database API}
\usage{
getWQ(stns, target_analytes = "all", 
matricesToExclude = "analyte_free_water",
output_colNames = c("stn", "date", "time", "param", "units", 
"matrix", "mdl", "value"),
output_colClasses = c("character", "character", "character", 
"character", "character", "character", "numeric", "numeric"),
rFriendlyParamNames = FALSE)
}
\arguments{
\item{stns}{a single-column dataframe of DataForEver station names. Case insensitive.}

\item{target_analytes}{grep-style character vector naming analytes of interest. default is "all". e.g., # target_analytes <- c("PHOSPHATE|NITROGEN|AMMONI|SUSPENDED|DISSOLVED OXYGEN|CALCIUM|POTASSIUM|HARDNESS|SODIUM|CHLORIDE|TEMP|CONDUCTIVITY, FIELD|SILICA|LEAD, TOTAL|MAGNESIUM|TURBIDITY|CHLOROPHYLL|MERCURY, TOTAL|SULFATE|ZINC, TOTAL|CHLORDANE|MALATHION|CARBOPHENOTHION|PH, FIELD")}

\item{matricesToExclude}{character vector specifying any sample matrices to be excluded. Spelling/case must be a perfect match to DataForEver entries. For example, "analyte_free_water" indicates field blanks. Advisible to check output using, e.g., \code{unique(wqDat$matrix)}}

\item{output_colNames}{names of columns in output (changing this is not recommended; this argument is more of a placeholder)}

\item{output_colClasses}{classes of columns in output (changing this is not recommended; this argument is more of a placeholder)}

\item{rFriendlyParamNames}{TRUE/FALSE; indicates whether parameter names should be modified to be R-friendly (no special characters, commas, or spaces). Advisable for analysis, as this makes analysis easier and pre-empts changes coerced by, e.g., \code{plyr::ddply}}
}
\value{
dataframe \code{getWQ} returns a dataframe with water quality measurements from each station identified in \code{stns}.
}
\description{
Downloads and compiles DataForEver water quality data. This function Works only on linux machines on the SFNRC network with access to the opt/physical drive. Code issues system commands, runs shell scripts, and modifies files in a temp folder on the local hard drive.
}
\examples{
\dontrun{
stations <- c("S333", "S12A", "S12B", "S12C", "S12D")

head(getWQ(stns = stations, target_analytes = c("PHOSPHATE|NITROGEN|AMMONI")))

### to generate the wqDat dataframe included in package:
targetStns <- c("AC01", "BB02", "BB04", "BB05A", "BB06", "BB09", "BB10", "BB11", 
"BB14", "BB15", "BB16", "BB17", "BB19", "BB22", "BB23", "BB24", 
"BB25", "BB26", "BB27", "BB28", "BB29", "BB31", "BB32", "BB34", 
"BB35", "BB36", "BB37", "BB38", "BB39A", "BB41", "BB42", "BB43", 
"BB44", "BB45", "BB47", "BB48", "BB52", "BB53", "BISC101", "BISC102", 
"BISC103", "BISC104", "BISC105", "BISC106", "BISC107", "BISC108", 
"BISC109", "BISC110", "BISC111", "BISC112", "BISC113", "BISC114", 
"BISC115", "BISC116", "BISC117", "BISC119", "BISC120", "BISC121", 
"BISC122", "BISC123", "BISC124", "BISC125", "BISC126", "BISC127", 
"BISC128", "BISC129", "BISC130", "BISC131", "BISC132", "BISC133", 
"BISC134", "BISC135", "BL01", "BS01", "LR01", "MI01", "TPBBSW-1", 
"TPBBSW-10", "TPBBSW-14", "TPBBSW-1B", "TPBBSW-1T", "TPBBSW-2", 
"TPBBSW-2B", "TPBBSW-3", "TPBBSW-3B", "TPBBSW-4", "TPBBSW-4B", 
"TPBBSW-4T", "TPBBSW-5", "TPBBSW-5B", "TPBBSW-5T", "TPGW-11D", 
"TPGW-11M", "TPGW-11S", "TPGW-14D", "TPGW-14M", "TPGW-14S", "UPKEYS201", 
"UPKEYS202", "UPKEYS203", "UPKEYS204", "UPKEYS205", "UPKEYS206", 
"UPKEYS207", "UPKEYS208", "UPKEYS210", "S20", "S20F", "S20G", 
"S21A", "S21", "S22", "S25", "S25A", "S25B", "S26", "S27", "S28", 
"G58", "S700", "G93", "S123", "S197")

wq.df <- getWQ(stns = targetStns)
}

}
\seealso{
\code{\link{getHydro}}
}
