% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydDat.R
\docType{data}
\name{hydDat}
\alias{hydDat}
\title{Data: DataForever hydrology data}
\format{A dataframe of flow, headwater, tailwater, and stage data for structures. Created on 19 October 2018.
\describe{
\item{stn}{station name}
\item{date}{date of record, POSIXct}
\item{flow}{flow (cfs)}
\item{head_water}{headwater stage (ft; NGVD29)}
\item{tail_water}{tailwater stage (ft; NGVD29)}
\item{stage}{stage (ft; NGVD29)}
\item{year}{year (extracted from date column)}
\item{mo}{month (extracted from date column)}
\item{day}{day (extracted from date column)}
\item{value}{amount of parameter measured}
}}
\usage{
hydDat
}
\description{
Data: DataForever hydrology data
}
\examples{
\dontrun{
### generated by: 
targetStns <- c("AC01", "BB02", "BB04", "BB05A", "BB06", "BB09", "BB10", "BB11", 
"BB14", "BB15", "BB16", "BB17", "BB19", "BB22", "BB23", "BB24", 
"BB25", "BB26", "BB27", "BB28", "BB29", "BB31", "BB32", "BB34", 
"BB35", "BB36", "BB37", "BB38", "BB39A", "BB41", "BB42", "BB43", 
"BB44", "BB45", "BB47", "BB48", "BB52", "BB53", "BISC101", "BISC102", 
"BISC103", "BISC104", "BISC105", "BISC106", "BISC107", "BISC108", 
"BISC109", "BISC110", "BISC111", "BISC112", "BISC113", "BISC114", 
"BISC115", "BISC116", "BISC117", "BISC119", "BISC120", "BISC121", 
"BISC122", "BISC123", "BISC124", "BISC125", "BISC126", "BISC127", 
"BISC128", "BISC129", "BISC130", "BISC131", "BISC132", "BISC133", 
"BISC134", "BISC135", "BL01", "BS01", "LR01", "MI01", "TPBBSW-1", 
"TPBBSW-10", "TPBBSW-14", "TPBBSW-1B", "TPBBSW-1T", "TPBBSW-2", 
"TPBBSW-2B", "TPBBSW-3", "TPBBSW-3B", "TPBBSW-4", "TPBBSW-4B", 
"TPBBSW-4T", "TPBBSW-5", "TPBBSW-5B", "TPBBSW-5T", "TPGW-11D", 
"TPGW-11M", "TPGW-11S", "TPGW-14D", "TPGW-14M", "TPGW-14S", "UPKEYS201", 
"UPKEYS202", "UPKEYS203", "UPKEYS204", "UPKEYS205", "UPKEYS206", 
"UPKEYS207", "UPKEYS208", "UPKEYS210", "S20", "S20F", "S20G", 
"S21A", "S21", "S22", "S25", "S25A", "S25B", "S26", "S27", "S28", 
"G58", "S700", "G93", "S123", "S197",
"S333", "S12A", "S12B", "S12C", "S12D", "S151")

hydDat <- getHydro(stns = targetStns, # c(as.character(dfrInBay$stn), structs), 
parameter_list = c("flow", "head_water", "salinity", "temperature", "tail_water", 
"stage", "rainfall", "precipitation", "ppt"), data_shape = "wide")

# save(hydDat, file = "/home/thill/RDATA/git-repos/SFNRC/data/hydDat.RData")

}

head(hydDat)

}
\keyword{bay,}
\keyword{biscayne}
\keyword{data}
\keyword{hydrology}
