% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialCoords.R
\name{makeSpatialCoords}
\alias{makeSpatialCoords}
\title{Make EDEN/DBHYDRO station coordinates spatial data}
\usage{
makeSpatialCoords(
  data,
  format = "dbhydro",
  new_crs = "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
)
}
\arguments{
\item{data}{input data. Typically, output from `getCoords_XXXX` functions.}

\item{format}{data source: 'eden' or 'dbhydro'}

\item{new_crs}{coordinate system to convert data to. Optional (set to `NULL` if a conversion is not desired).}
}
\value{
spatialPointsDataFrame converts the output from getCoords_EDEN and getCoords_DBHYDRO to a spatialPointsDataFrame. EDEN data are set in the same projection as the EDEN DEM ("+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs") and DBHYDRO data are in lat/long ("+init=epsg:4326")
}
\description{
Converts scraped EDEN/DBHYDRO station coordinates to spatial data.
}
\examples{
\dontrun{
### example usage as part of a workflow

stns <- c(
"S18C_T",
"NESRS1"
)
stn.coords <- do.call(rbind, lapply(X = stns, getCoords_EDEN))

### convert to spatial data and plot
stn.coords <- makeSpatialCoords(stn.coords, format = "eden")

plot(stn.coords)
}

}
