% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterCoords.R
\docType{data}
\name{masterCoords}
\alias{masterCoords}
\title{Data: Coordinates for water quality stations in DataForEver}
\format{A dataframe with 3562 water quality stations:
\describe{
\item{stn}{station name}
\item{long}{station longitude in decimal degrees}
\item{lat}{station latitude in decimal degrees}
}}
\usage{
masterCoords
}
\description{
Data: Coordinates for water quality stations in DataForEver
}
\examples{
### export to .csv:
write.csv(masterCoords, file = file.path(tempdir(), "masterCoords.csv"))
}
\keyword{data,}
\keyword{quality}
\keyword{water}
