% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculates standard error}
\usage{
se(x)
}
\arguments{
\item{x}{numeric or integer}
}
\value{
value
}
\description{
Calculates standard error
}
\examples{
se(rnorm(10, 5, 5))
se(rnorm(100, 5, 5))

}
