% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversion.R
\name{unitConversion}
\alias{unitConversion}
\title{Convert units}
\usage{
unitConversion(
  data.source = "dss",
  station.type = "transect",
  units = "kaf",
  graph_type = "barchart",
  digits.precision = 3,
  DEBUG = FALSE
)
}
\arguments{
\item{data.source}{can be "dss" or ...}

\item{station.type}{station type. options: transect}

\item{units}{units desired. options: kaf, cfs, cms}

\item{graph_type}{graph produced}

\item{digits.precision}{decimal places reported}

\item{DEBUG}{not sure what this is for}
}
\value{
value
}
\description{
Convert units
}
