% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqDat.R
\docType{data}
\name{wqDat}
\alias{wqDat}
\title{Data: DataForEver water quality data for northern boundary of Everglades National Park}
\format{A dataframe with x observations of y variables:
\describe{
\item{stn}{station name}
\item{date}{date of observation}
\item{param}{parameter name}
\item{units}{measurement units}
\item{year}{year (extracted from date column)}
\item{value}{amount of parameter measured}
\item{mdl}{method detection limit, where available}
\item{src}{data source (DataForEver)}
}}
\usage{
wqDat
}
\description{
Data: DataForEver water quality data for northern boundary of Everglades National Park
}
\examples{
### export to .csv:
write.csv(wqDat, file = file.path(tempdir(), "wqDat.csv"))
}
\keyword{data,}
\keyword{quality}
\keyword{water}
