% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnp.R
\docType{data}
\name{bnp}
\alias{bnp}
\title{Data: Biscayne Bay political boundaries}
\format{A shapefile of political boundaries in Biscayne Bay, Florida. These are a subset of areas of Biscayne Bay defined as Outstanding Florida Waters. Coordinate reference system is "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
\describe{
\item{OBJECTID}{a SpatialPolygonDataFrame of Biscayne Bay}
\item{CODE}{area of basin in hectares}
\item{LANDTYPE}{expanded name of polygon zone}
\item{CODE2}{unclear}
\item{ACRES}{area of polygon in acres}
\item{SFRPC_CODE}{unclear}
\item{AMERICODE}{unclear}
\item{WATER_SHD}{unclear}
\item{GROUP_SHD}{unclear}
\item{WATERSHED}{unclear}
\item{Shape_Leng}{unclear}
\item{Shape_Area}{area of polygon, units unclear}
}}
\usage{
bnp
}
\description{
Data: Biscayne Bay political boundaries
}
\examples{
plot(bnp)

\dontrun{
### code used to generate object
shp <- "/opt/physical/troy/ESRIDATA/enp_boundary_line/PublicLandsBoundary.shp" # "/opt/physical/troy/ESRIDATA/BNP_boundary.shp"
bnp <- readOGR(shp, layer = basename(strsplit(shp, "\\\\.")[[1]])[1], pointDropZ = TRUE, stringsAsFactors = FALSE)
bnp <- spTransform(bnp, CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))
### isolate just the areas of interest
bnp <- bnp[as.character(bnp$OBJECTID) \%in\% c("423", "363", "290", "192"), ]
}

}
\keyword{bay,}
\keyword{biscayne}
\keyword{map}
