% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_proj_batch.R
\name{dbhydro.proj.batch}
\alias{dbhydro.proj.batch}
\title{A wrapper for \code{dbhydro.proj}}
\usage{
dbhydro.proj.batch(codes, ...)
}
\arguments{
\item{codes}{a character vector of DBHYDRO project names.}

\item{...}{additional arguments supplied to \link{\code{dbhydro.proj}}}
}
\value{
dataframe \code{dbhydro.proj.batch} saves a csv of DBHYDRO water quality data to disk, with option to import data into global environment.
}
\description{
Downloads DBHYDRO water quality data for specified projects. Avoids the problem of complete failure if a single project is invalid.
}
\examples{
\dontrun{
STA1W_project_codes <- c("ENRM", "ENRP", "ENRU") 

dbh.proj.batch(codes = STA1W_project_codes)
}

}
\seealso{
\code{dbhydro.proj}
}
