% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_stn_batch.R
\name{dbhydro.stn.batch}
\alias{dbhydro.stn.batch}
\title{A wrapper for \code{dbhydro.stn}}
\usage{
dbhydro.stn.batch(codes, 
destfile = "stn_report_todaysDate.csv",
rename_proj = TRUE,
parameters = "all",
report_type = "full", # full = long dataset - 1 line per sample; crosstab = wide dataset
incl_qc_flags = TRUE,
incl_flagged_data = FALSE,
destination = "file_csv", 
start_date = "01-JAN-1960", # note format
end_date   = "today",
import_data = FALSE)
}
\arguments{
\item{codes}{a character vector of station names. If one station in the list returns no data, the entire url request fails.}

\item{destfile}{a character vector file address for output, saved as a comma-delimited text file.}

\item{rename_proj}{lorem ipsum}

\item{parameters}{lorem ipsum}

\item{report_type}{lorem ipsum}

\item{incl_qc_flags}{lorem ipsum}

\item{incl_flagged_data}{lorem ipsum}

\item{destination}{lorem ipsum}

\item{start_date}{lorem ipsum}

\item{end_date}{lorem ipsum}

\item{import_data}{lorem ipsum}
}
\value{
dataframe \code{dbhydro.stn.batch} saves a csv of DBHYDRO water quality data to disk.
}
\description{
Downloads DBHYDRO water quality data for specified stations. Avoids the problem of complete failure if a single station is invalid.
}
\examples{
\dontrun{
head(dbh.stn.batch())
}

}
\seealso{
\code{dbhydro.stn}
}
