% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_hydro.R
\name{dfe.hydro}
\alias{dfe.hydro}
\title{DataForEver hydrology database API}
\usage{
dfe.hydro(stns, parameter_list = c("flow", "tail_water", "head_water", "stage"),
data_shape = "long")
}
\arguments{
\item{stns}{a single-column dataframe of DataForEver station names.}

\item{parameter_list}{vector of desired parameters.}

\item{data_shape}{shape of output dataframe. Default is \code{long} (one row per date-station-param) but can also be \code{wide} (one row per date-station) or \code{really_wide} (one row per date)}
}
\value{
dataframe \code{dfe.hydro} returns a dataframe with water quality measurements from each station identified in \code{stns}.
}
\description{
Downloads and compiles DataForEver hydrology data
}
\details{
This function Works only on linux machines on the SFNRC network with access to the opt/physical drive. Code issues system commands, runs shell scripts, and modifies files in a temp folder on the local hard drive.
}
\examples{
\dontrun{
stations <- c("S333", "S12A", "S12B", "S12C", "S12D")

  # usage examples: 
  hydro.test <- dfe.hydro(stns = stations)
  hydro.test2 <- dfe.hydro(stns = stations, data_shape = "wide")
  hydro.test3 <- dfe.hydro(stns = stations, data_shape = "really_wide")
}

}
\seealso{
\code{\link{dfe.wq}}
}
