% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finDat.R
\docType{data}
\name{finDat}
\alias{finDat}
\title{Data: Merged DERM and DataForever water quality data}
\format{A dataframe with 792057 observations of 7 variables. Created on 31 May 2019.
\describe{
\item{stn}{station name}
\item{date}{date of observation}
\item{param}{parameter name}
\item{units}{measurement units}
\item{year}{year (extracted from date column)}
\item{value}{amount of parameter measured}
\item{mdl}{method detection limit, where available}
}}
\usage{
finDat
}
\description{
Data: Merged DERM and DataForever water quality data
}
\examples{
head(finDat)
### finDat generated by opt/physical/troy/RDATA/biscayne/scripts/script_dermDFE_prep_20181019.R

}
\keyword{bay,}
\keyword{biscayne}
\keyword{data}
