% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordsEDEN.R
\name{getCoords_EDEN}
\alias{getCoords_EDEN}
\title{Retrieve station coordinates from EDEN}
\usage{
getCoords_EDEN(stn = "S18C_T")
}
\arguments{
\item{stn}{character string. Case sensitive.}
}
\value{
dataframe \code{getCoords_EDEN} returns a dataframe with lat/long (epsg:4326) UTM zone 19 coordinates, and the NAVD-to-NGVD conversion factor (units = feet). [ft. NAVD88] + [conversion factor] = [ft. NGVD29]
}
\description{
Scrapes EDEN station coordinate data.
}
\examples{
\dontrun{
### example usage as part of a workflow

stns <- c(
"S18C_T",
"NESRS1"
)
stn.coords <- do.call(rbind, lapply(X = stns, getCoords_EDEN))

### convert to spatial data and plot
stn.coords           <- stn.coords[!is.na(stn.coords$latitude), ]
stn.coords$longitude <- -1 * stn.coords$longitude
coordinates(stn.coords) <- c("easting", "northing")
proj4string(stn.coords) <- "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs" # crs(fireHydro::edenDEM)

plot(stn.coords)
}

}
