% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBHYDRO.R
\name{getDBHYDRO}
\alias{getDBHYDRO}
\title{Produce R-friendly water quality data from DBHYDRO}
\usage{
getDBHYDRO(stn = "S333", parameters = c("PHOSPHATE, TOTAL AS P", "TURBIDITY"), 
    removeFlaggedData = TRUE, outputType = "wide")
}
\arguments{
\item{stn}{character string. Case insensitive.}

\item{parameters}{water quality parameters to report. Set to NA to report all available parameters.}

\item{removeFlaggedData}{Logical. Indicates whether flagged data should be removed}

\item{outputType}{output type can be "full", "long", or "wide". "full" is a long dataset with all samples (possibility of multiple samples on a single day). "long" and "wide" average duplicate samples and are either long datasets (one column with parameter names and one column with parameter values) or wide datasets (one column of values for each parameter). The "long" form reports more information (units, MDL, PQL, RDL) than the "wide" form, which only reports values for each parameter.}
}
\value{
dataframe \code{getDBHYDRO} returns a dataframe of data
}
\description{
Downloads and lightly processes DBHYDRO data to make it easier to work with in R.
}
\examples{
\dontrun{
longDat <- getDBHYDRO(stn = "s333", outputType = "long")
wideDat <- getDBHYDRO(stn = "s333", outputType = "wide")
}

}
