% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBHYDROhydro.R
\name{getDBHYDROhydro}
\alias{getDBHYDROhydro}
\title{Retrieve R-friendly hydrology data from DBHYDRO using DB keys}
\usage{
getDBHYDROhydro(dbkey = "03638", startDate = "19600101",
endDate = gsub(x = Sys.Date(), pattern = "-", replacement = ""))
}
\arguments{
\item{dbkey}{DBkey for station and parameter of interest. Use 'getDBkey()' to find DBkeys associated with a station.}

\item{startDate}{beginning of period of record, in form YYYYMMDD}

\item{endDate}{end of period of record, in form YYYYMMDD}
}
\value{
dataframe \code{getDBHYDROhydro} returns a dataframe of data
}
\description{
Downloads and lightly processes DBHYDRO data. This function uses DB keys to download datasets and is therefore more versatile than getDBHYDRO.
}
\examples{
\dontrun{
### example workflow:
getDBkey(stn = "S333", type = "FLOW") # find DBkey for station/parameter of interest
s333.daily <- getDBHYDROhydro(dbkey = "15042") # download data for DBkey
tail(s333.daily)

### check that instantaneous data is also loaded properly
### downloading large datasets may take a very long time
s333.inst <- getDBHYDROhydro(dbkey = "65086", 
   startDate = gsub(x = Sys.Date(), pattern = "-", replacement = ""))
tail(s333.inst)
}

}
