% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBHYDROhydro.R
\name{getDBHYDROhydro}
\alias{getDBHYDROhydro}
\title{Retrieve R-friendly hydrology data from DBHYDRO using DB keys}
\usage{
getDBHYDROhydro(dbkey = "03638", startDate = "19600101",
endDate = gsub(x = Sys.Date(), pattern = "-", replacement = ""))
}
\arguments{
\item{dbkey}{DBkey for station and parameter of interest. Use 'getDBkey()' to find DBkeys associated with a station.}

\item{startDate}{beginning of period of record, in form YYYYMMDD}

\item{endDate}{end of period of record, in form YYYYMMDD}
}
\value{
dataframe \code{getDBHYDROhydro} returns a dataframe of data
}
\description{
Downloads and lightly processes DBHYDRO data. This function uses DB keys to download datasets and is therefore more versatile than getDBHYDRO.
}
\examples{
\dontrun{
### example workflow:
getDBkey(stn = "S12d", type = "FLOW") # find DBkey for station/parameter of interest
fdat <- getDBHYDROhydro(dbkey = "01310") # download data for DBkey
tail(fdat)

### check that instantaneous data is also loaded properly
s333.inst <- getDBHYDROhydro(dbkey = "65086")
tail(s333.inst)
}

}
