% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_data_types.R
\name{getParams_DFE}
\alias{getParams_DFE}
\title{DataForEver hydrology database parameter query}
\usage{
getParams_DFE(dbname = "hydrology", stn = "all")
}
\arguments{
\item{dbname}{name of the database sought for inquiry. Currently only 'hydrology' and 'waterquality' are supported. A case-insensitive character string.}

\item{stn}{pattern to be matched in station names ('NULL' or 'all' return all stations). This argument applies only to water quality database. A case-insensitive grep-friendly single character element (e.g., 'S333|S197' to search for multiple stations).}
}
\value{
dataframe \code{getParams_DFE} returns a dataframe of stations and parameters.
}
\description{
Identifies parameters available from the DataForEver hydrology database. Note that the structure of the returned dataframe can vary depending on the database queried.
}
\examples{
\dontrun{
### station argument applies only to water quality database
t1 <- getParams_DFE(stn = "all", dbname = "hydrology")       # all hydrology parameters
t2 <- getParams_DFE(stn = "all", dbname = "waterquality")    # all water quality parameters
t3 <- getParams_DFE(stn = "S333|S18C", dbname = "hydrology") # station-specific hydro report 
t4 <- getParams_DFE(stn = "S333|S18C", dbname = "waterquality") # station-specific WQreport
}


}
