% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derm_wq.R
\name{procDERM}
\alias{procDERM}
\title{Process water quality data from Miami-Dade Department of Environmental Resources Management}
\usage{
procDERM(filename = "/opt/physical/troy/RDATA/biscayne/data/data_BBWQall_20180517.txt",
remove_QC_codes = c("B|H|J|K|Q|Y|\\\\?|\\\\*"), 
new_names = c("stn", "date", "time", "param", "method", "units", 
"QCcode", "mdl", "value", "dilution.factor", "depth", "depth.code", 
"x", "y", "matrix", "labComments", "project", "genComments", "lab", 
"year", "mo", "day", "datetime"))
}
\arguments{
\item{filename}{a character vector file address for DERM data, saved as a tab-delimited text file.}

\item{remove_QC_codes}{Quality control codes to remove from the dataset.}

\item{new_names}{names of columns in output (changing this is not recommended)}
}
\value{
dataframe \code{procDERM} returns a dataframe with water quality measurements, with parameter names and units that match DataForEver data.
}
\description{
Processes DERM data to facilitate analysis and integration with DataForEver data. Adjusts parameter names and units to match DataForEver.
}
\details{
This function is tailored specifically for DERM datasets.
}
\examples{
\dontrun{
head(procDERM())
}

}
\seealso{
\code{none}
}
