% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas.R
\name{seas}
\alias{seas}
\title{Create columns indicating wet/dry season and federal water year}
\usage{
seas(inputData, timeCol = "datetime", wetSeas = c("May", "Sep"), waterYearBegin = "Oct")
}
\arguments{
\item{inputData}{input dataframe}

\item{timeCol}{POSIXct column of timestamps}

\item{wetSeas}{the first and last month of the wet season (inclusive; use abbreviated months as in \code{month.abb})}

\item{waterYearBegin}{the first full month of each water year (abbreviated form, as in \code{month.abb}).}
}
\value{
two columns are appended to the input dataframe
}
\description{
See above. Water year 2015 runs from Oct 01 2014 - Sept 30 2015.
}
\examples{
a <- seas(wqDat, timeCol = "datetime")
head(a)

}
